last? conversions for #139
