+++ unittests/unordered_set.cpp
@@ -24,110 +24,29 @@
   (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
   SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
-#include "common.hpp"
-#include <boost/test/unit_test.hpp>
+#define DOCTEST_CONFIG_IMPLEMENT_WITH_MAIN
+#include "unordered_set.hpp"
 
-template <class IArchive, class OArchive>
-void test_unordered_set()
-{
-  std::random_device rd;
-  std::mt19937 gen(rd());
+TEST_SUITE("unordered_set");
 
-  for(int ii=0; ii<100; ++ii)
-  {
-    std::unordered_set<int> o_podunordered_set;
-    for(int j=0; j<100; ++j)
-      o_podunordered_set.insert(random_value<int>(gen));
-
-    std::unordered_set<StructInternalSerialize, StructHash<StructInternalSerialize>> o_iserunordered_set;
-    for(int j=0; j<100; ++j)
-      o_iserunordered_set.insert({ random_value<int>(gen), random_value<int>(gen) });
-
-    std::unordered_set<StructInternalSplit, StructHash<StructInternalSplit>> o_isplunordered_set;
-    for(int j=0; j<100; ++j)
-      o_isplunordered_set.insert({ random_value<int>(gen), random_value<int>(gen) });
-
-    std::unordered_set<StructExternalSerialize, StructHash<StructExternalSerialize>> o_eserunordered_set;
-    for(int j=0; j<100; ++j)
-      o_eserunordered_set.insert({ random_value<int>(gen), random_value<int>(gen) });
-
-    std::unordered_set<StructExternalSplit, StructHash<StructExternalSplit>> o_esplunordered_set;
-    for(int j=0; j<100; ++j)
-      o_esplunordered_set.insert({ random_value<int>(gen), random_value<int>(gen) });
-
-    std::ostringstream os;
-    {
-      OArchive oar(os);
-
-      oar(o_podunordered_set);
-      oar(o_iserunordered_set);
-      oar(o_isplunordered_set);
-      oar(o_eserunordered_set);
-      oar(o_esplunordered_set);
-    }
-
-    std::unordered_set<int> i_podunordered_set;
-    std::unordered_set<StructInternalSerialize, StructHash<StructInternalSerialize>>   i_iserunordered_set;
-    std::unordered_set<StructInternalSplit, StructHash<StructInternalSplit>>           i_isplunordered_set;
-    std::unordered_set<StructExternalSerialize, StructHash<StructExternalSerialize>>   i_eserunordered_set;
-    std::unordered_set<StructExternalSplit, StructHash<StructExternalSplit>>           i_esplunordered_set;
-
-    std::istringstream is(os.str());
-    {
-      IArchive iar(is);
-
-      iar(i_podunordered_set);
-      iar(i_iserunordered_set);
-      iar(i_isplunordered_set);
-      iar(i_eserunordered_set);
-      iar(i_esplunordered_set);
-    }
-
-    for(auto const & p : i_podunordered_set)
-    {
-      BOOST_CHECK_EQUAL(o_podunordered_set.count(p), 1lu);
-    }
-
-    for(auto const & p : i_iserunordered_set)
-    {
-      BOOST_CHECK_EQUAL(o_iserunordered_set.count(p), 1lu);
-    }
-
-    for(auto const & p : i_isplunordered_set)
-    {
-      BOOST_CHECK_EQUAL(o_isplunordered_set.count(p), 1lu);
-    }
-
-    for(auto const & p : i_eserunordered_set)
-    {
-      BOOST_CHECK_EQUAL(o_eserunordered_set.count(p), 1lu);
-    }
-
-    for(auto const & p : i_esplunordered_set)
-    {
-      BOOST_CHECK_EQUAL(o_esplunordered_set.count(p), 1lu);
-    }
-  }
-}
-
-BOOST_AUTO_TEST_CASE( binary_unordered_set )
+TEST_CASE("binary_unordered_set")
 {
   test_unordered_set<cereal::BinaryInputArchive, cereal::BinaryOutputArchive>();
 }
 
-BOOST_AUTO_TEST_CASE( portable_binary_unordered_set )
+TEST_CASE("portable_binary_unordered_set")
 {
   test_unordered_set<cereal::PortableBinaryInputArchive, cereal::PortableBinaryOutputArchive>();
 }
 
-BOOST_AUTO_TEST_CASE( xml_unordered_set )
+TEST_CASE("xml_unordered_set")
 {
   test_unordered_set<cereal::XMLInputArchive, cereal::XMLOutputArchive>();
 }
 
-BOOST_AUTO_TEST_CASE( json_unordered_set )
+TEST_CASE("json_unordered_set")
 {
   test_unordered_set<cereal::JSONInputArchive, cereal::JSONOutputArchive>();
 }
 
-
+TEST_SUITE_END();
